# openapi.yaml
openapi: 3.0.0
info:
  title: 'Cube-3 API'
  contact:
    email: leo.paillard@gmail.com
  version: 1.0.0
servers:
  -
    url: 'http://cyber-dodo.fr:8080/api/v1'
    description: 'Cube-3 API'
paths:
...
  /project-status:
    post:
      tags:
        - ProjectStatus
      summary: 'Add a new ProjectStatus'
      description: 'Add a new ProjectStatus'
      operationId: 0fefaf16e5acb18be8768e515d6a8a1f
      requestBody:
        description: 'ProjectStatus object that needs to be added'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectStatusRequest'
      responses:
        '201':
          description: 'ProjectStatus created'
        '400':
          description: 'Invalid request data'
        '409':
          description: 'ProjectStatus already exists'
        '500':
          description: 'Internal Server Error'
  /project-status/all:
    get:
      tags:
        - ProjectStatus
      summary: 'Get all ProjectStatuses'
      description: 'Get all ProjectStatuses'
      operationId: 1eedc4528a3f9a2f7b748a723e49eb08
      responses:
        '200':
          description: 'ProjectStatuses found'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectStatusResponse'
        '500':
          description: 'Internal Server Error'
  '/project-status/{id}':
    get:
      tags:
        - ProjectStatus
      summary: 'Get a ProjectStatus by ID'
      description: 'Get a ProjectStatus by ID'
      operationId: 715c693baa7822fc4cd8026476b98fe8
      parameters:
        -
          name: id
          in: path
          description: 'ID of the ProjectStatus to get'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'ProjectStatus found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectStatusResponse'
        '404':
          description: 'ProjectStatus not found'
        '500':
          description: 'Internal Server Error'
    put:
      tags:
        - ProjectStatus
      summary: 'Update a ProjectStatus by ID'
      description: 'Update a ProjectStatus by ID'
      operationId: dcfdc5fc41489e0ed11ce0432359ed5f
      parameters:
        -
          name: id
          in: path
          description: 'ID of the ProjectStatus to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: 'ProjectStatus object that needs to be updated'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectStatusRequest'
      responses:
        '200':
          description: 'ProjectStatus updated'
        '400':
          description: 'Invalid request data'
        '404':
          description: 'ProjectStatus not found'
        '409':
          description: 'ProjectStatus already exists'
        '500':
          description: 'Internal Server Error'
    delete:
      tags:
        - ProjectStatus
      summary: 'Delete a ProjectStatus by ID'
      description: 'Delete a ProjectStatus by ID'
      operationId: b82cc4646d77534aadd935052e28e065
      parameters:
        -
          name: id
          in: path
          description: 'ID of the ProjectStatus to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'ProjectStatus deleted'
...
components:
  schemas:
...
    ProjectStatusRequest:
      required:
        - name
        - description
      properties:
        name:
          type: string
          example: 'ProjectStatus 1'
        description:
          type: string
          example: 'This is the first projectStatus'
      type: object
    ProjectStatusResponse:
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'ProjectStatus 1'
        description:
          type: string
          example: 'This is the first projectStatus'
        createdAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2021-01-01 00:00:00'
      type: object
...
  securitySchemes:
    bearer:
      type: http
      bearerFormat: JWT